+++ test/suites/dev-tanssi-relay/test-basics.ts
@@ -26,11 +26,14 @@ describeSuite({
             title: "Checking that launched node can create blocks",
             test: async function () {
                 const block = (await polkadotJs.rpc.chain.getBlock()).block.header.number.toNumber();
-                await context.createBlock();
+
+                for (let i = 0; i < 12; i++) {
+                    await context.createBlock();
+                }
 
                 const block2 = (await polkadotJs.rpc.chain.getBlock()).block.header.number.toNumber();
                 log(`Original block #${block}, new block #${block2}`);
-                expect(block2).to.be.greaterThan(block);
+                expect(block2).to.be.equal(block + 12);
             },
         });
 
@@ -40,13 +43,12 @@ describeSuite({
             timeout: 200000,
             test: async function () {
                 const balanceBefore = (await polkadotJs.query.system.account(bob.address)).data.free.toBigInt();
-
-                await polkadotJs.tx.balances.transferAllowDeath(bob.address, 1000).signAndSend(charlie);
-
-                await context.createBlock();
+                const tx = await polkadotJs.tx.balances
+                    .transferAllowDeath(bob.address, 100000000000000n)
+                    .signAsync(charlie);
+                await context.createBlock([tx]);
                 const balanceAfter = (await polkadotJs.query.system.account(bob.address)).data.free.toBigInt();
-
-                expect(balanceAfter).to.be.equal(balanceBefore + 1000n);
+                expect(balanceAfter).to.be.equal(balanceBefore + 100000000000000n);
             },
         });
 
@@ -56,10 +58,8 @@ describeSuite({
             timeout: 200000,
             test: async function () {
                 await context.createBlock();
-                //const tx = polkadotJs.tx.rootTesting.fillBlock(60 * 10 ** 7);
-                await polkadotJs.tx.sudo
-                    .sudo(polkadotJs.tx.balances.transferAllowDeath(bob.address, 1000))
-                    .signAndSend(alice);
+                const tx = polkadotJs.tx.rootTesting.fillBlock(60 * 10 ** 7);
+                await polkadotJs.tx.sudo.sudo(tx).signAndSend(alice);
 
                 await context.createBlock();
                 const blockFill = await polkadotJs.query.system.blockWeight();
